requirements.txt
fastapi[standard]~=0.115.6
gql~=3.5.0
typer~=0.15.1
aiohttp~=3.11.11
pymongo~=4.10.1
PyGithub~=2.5.0
celery~=5.4.0
pymediator~=0.3.0
pydantic~=2.10.4
starlette~=0.41.3
pydantic-settings~=2.7.1
python-dotenv~=1.0.1